fix : 페이지 로딩이 되지않는 오류 수정 및 Input칸 입력 수 제한, 리다이렉트 처리 활성화
