--- src/RequireAuth.tsx
@@ -1,20 +1,0 @@
-import { Navigate } from 'react-router-dom';
-import useUserStore, { type userStore } from './stores/useUserStore';
-
-export default function RequireAuth({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  const user = useUserStore((store: userStore) => store.user);
-
-  if (!user) {
-    return <Navigate to='/onboarding' replace />;
-  }
-
-  if (user && window.location.pathname === '/onboarding') {
-    return <Navigate to='/' replace />;
-  }
-
-  return <>{children}</>;
-}
