+++ src/routes.tsx
@@ -1,6 +1,6 @@
 import { lazy } from 'react';
 import { createBrowserRouter, Outlet } from 'react-router-dom';
-// import RequireAuth from './RequireAuth.tsx';
+import RequireAuth from './utils/RequireAuth.tsx';
 
 const Layout = lazy(() => import('./components/layout/Layout'));
 const HomePage = lazy(() => import('./pages/HomePage'));
@@ -29,20 +29,18 @@ const router = createBrowserRouter([
         index: true,
         element: (
           <BasicLayout>
-            {/* <RequireAuth> */}
-            <HomePage />
-            {/* </RequireAuth> */}
+            <RequireAuth>
+              <HomePage />
+            </RequireAuth>
           </BasicLayout>
         ),
-        // 이후 리다이렉트 설정할 때 조건문과 함께 아래 루트 수정
-        // element: <Navigate to='onboarding' replace />,
       },
       {
         path: '/onboarding',
         element: (
-          // <RequireAuth>
-          <OnboardingPage />
-          // </RequireAuth>
+          <RequireAuth>
+            <OnboardingPage />
+          </RequireAuth>
         ),
       },
       {
